bugs fixed

+++ lib/Models/StaffModel.dart
@@ -12,7 +12,7 @@ class StaffModel{
   late String userType;
   final String deptName;
   late String firstSignin;
-  int casualLeaveTaken;
+  double casualLeaveTaken;
 
   StaffModel({
     required this.fullName,
@@ -62,5 +62,4 @@ class StaffModel{
       casualLeaveTaken: snapShot['casualleavetaken']
     );
   }
-
 }
